pyproject.toml
[project]
name = "imap-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioimaplib>=2.0.1",
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
    "msal>=1.34.0",
    "oauthlib>=3.3.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "streamlit>=1.51.0",
]
